780:adding forward icon in the article bellow,and extract same code to one component
