+++ frontends/article/article-page.tsx
@@ -1,17 +1,9 @@
 import EditorJS from '@editorjs/editorjs';
-import { Divider, Editor } from 'components';
-import { ApiError, Blog, Button, Head, StorysetImage } from 'components';
-import { useAuth, useTheme } from 'hooks';
-import Image from 'next/image';
+import { ApiError, Blog, Button, Divider, Editor, Head, StorysetImage } from 'components';
 import Link from 'next/link';
 import { FC, useEffect, useMemo, useState } from 'react';
 import { useAppDispatch } from 'store';
-import {
-  useIncrementViewCountMutation,
-  useLazyCheckIfLikedDislikedQuery,
-  useLikeDislikeMutation,
-} from 'store/apis';
-import { openLoginModal, toggleCommentsSidebar } from 'store/states';
+import { useIncrementViewCountMutation } from 'store/apis';
 import { setArticleAuthor } from 'store/states/readArticle';
 import { IArticle } from 'types';
 import {
@@ -22,15 +14,12 @@ import {
   get,
   toDateString,
 } from 'utils';
-import { ICONS, UPPER_BLUE_COLOR } from 'variables';
+import { ICONS } from 'variables';
 
 import styles from './article.module.scss';
 import { IArticleProps } from './article.types';
-import { ArticleActions } from './components';
+import { ArticleActionIcons, ArticleActions } from './components';
 
-const LikeIcon = ICONS.like;
-const DislikeIcon = ICONS.dislike;
-const CommentIcon = ICONS.comment;
 const HeartIcon = ICONS.heart;
 
 const toUzbDateString = (date: Date | string): string => toDateString(date, { month: 'short' });
@@ -42,39 +31,12 @@ export const Article: FC<IArticleProps> = ({
   showAuthor = false,
   ...props
 }) => {
-  const {
-    viewCount = 0,
-    publishedDate,
-    updatedDate,
-    blocks = [],
-    id,
-    likeCount = 0,
-    dislikeCount = 0,
-  } = article || {};
+  const { viewCount = 0, publishedDate, updatedDate, blocks = [] } = article || {};
   const [editorInstance, setEditorInstance] = useState<EditorJS | null>(null);
-  const [likeDislikeCount, setLikeDislikeCount] = useState<number>(likeCount - dislikeCount);
-  const { isAuthenticated } = useAuth();
-  const { themeColors } = useTheme();
+  const [isSharePopupOpen, setIsSharePopupOpen] = useState<boolean>(false);
+
   const dispatch = useAppDispatch();
   const [incrementViewCountRequest] = useIncrementViewCountMutation();
-  const [likeDislikeArticle, likeDislikeRes] = useLikeDislikeMutation();
-  const [checkIfLikedDislikedQuery, { data: isLikedOrDisliked }] =
-    useLazyCheckIfLikedDislikedQuery();
-
-  const likeDislike = (value: -1 | 1): void => {
-    if (!isAuthenticated) {
-      dispatch(openLoginModal());
-      return;
-    }
-    if (likeDislikeRes.isLoading || value === isLikedOrDisliked || !id) return;
-    likeDislikeArticle({ id, value }).then(() => {
-      setLikeDislikeCount((prev) => prev + value - (isLikedOrDisliked || 0));
-    });
-  };
-
-  const commentIconClickHandler = (): void => {
-    dispatch(toggleCommentsSidebar());
-  };
 
   useEffect(() => {
     if (editorInstance?.isReady) {
@@ -83,13 +45,6 @@ export const Article: FC<IArticleProps> = ({
       main?.scrollTo(0, 0);
     }
   }, [editorInstance?.isReady]);
-
-  useEffect(() => {
-    if (isAuthenticated && id) {
-      checkIfLikedDislikedQuery(id);
-      setLikeDislikeCount(likeCount - dislikeCount);
-    }
-  }, [isAuthenticated, id]);
 
   useEffect(() => {
     editorInstance?.render?.({ blocks: addUriToImageBlocks(blocks) });
@@ -123,17 +78,6 @@ export const Article: FC<IArticleProps> = ({
     if (publishedDate) return <span>{toUzbDateString(publishedDate)}</span>;
     return <></>;
   }, [publishedDate, updatedDate]);
-
-  const likeIcon = useMemo((): JSX.Element => {
-    if (isLikedOrDisliked === 0) {
-      return (
-        <div style={{ transform: 'rotate(180deg)', display: 'flex' }}>
-          <Image width={30} height={30} src='/icons/dislike.webp' />
-        </div>
-      );
-    }
-    return <LikeIcon color={isLikedOrDisliked === 1 ? UPPER_BLUE_COLOR : themeColors.icon} />;
-  }, [isLikedOrDisliked, themeColors]);
 
   if (!article) {
     if (error?.status === 500) return <ApiError className='container mt-2' error={error} />;
@@ -189,38 +133,16 @@ export const Article: FC<IArticleProps> = ({
             {dateContent}
           </div>
           <div className={styles.reactions}>
-            <div className={styles.reactionButtons}>
-              <span
-                data-action='open-comments'
-                className='pointer me-2'
-                onClick={commentIconClickHandler}
-              >
-                <CommentIcon color={themeColors.icon} />
-              </span>
-              <span
-                className={`pointer icon me-2 ${isLikedOrDisliked === 1 && 'icon--active'}`}
-                onClick={(): void => likeDislike(1)}
-              >
-                {likeIcon}
-              </span>
-              {Boolean(likeDislikeCount) && <span className='me-2'>{likeDislikeCount}</span>}
-              <span
-                className={`pointer icon ${isLikedOrDisliked === -1 && 'icon--active'}`}
-                onClick={(): void => likeDislike(-1)}
-              >
-                <DislikeIcon
-                  color={isLikedOrDisliked === -1 ? UPPER_BLUE_COLOR : themeColors.icon}
-                />
-              </span>
-            </div>
+            <ArticleActionIcons
+              right={0}
+              popupId='articleDetail'
+              isSharePopupOpen={isSharePopupOpen}
+              setIsSharePopupOpen={setIsSharePopupOpen}
+              article={article}
+            />
           </div>
         </div>
-        <ArticleActions
-          editor={editorInstance}
-          likeDislike={likeDislike}
-          isLikedOrDisliked={isLikedOrDisliked}
-          likeDislikeCount={likeDislikeCount}
-        />
+        <ArticleActions editor={editorInstance} article={article} />
       </div>
     </div>
   );
