+++ frontends/article/article.module.scss
@@ -9,14 +9,11 @@
     align-items: center;
 
     .reactions {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-
-      .reactionButtons {
-        display: flex;
-        align-items: center;
-      }
+      transform: translateY(0);
+      opacity: 1;
+      visibility: visible;
+      z-index: 1999;
+      position: relative;
     }
 
     .navigation {
