+++ frontends/article/components/ArticleActions/article-actions.types.ts
@@ -1,8 +1,16 @@
 import EditorJS from '@editorjs/editorjs';
+import { Dispatch, SetStateAction } from 'react';
+import { IArticle } from 'types';
 
 export interface IArticleActionsProps {
   editor: EditorJS | null;
-  isLikedOrDisliked?: number;
-  likeDislikeCount: number;
-  likeDislike: (value: -1 | 1) => void;
+  article: IArticle | null;
+}
+
+export interface IArticleActionsIconsProps {
+  article: IArticle | null;
+  isSharePopupOpen: boolean;
+  setIsSharePopupOpen: Dispatch<SetStateAction<boolean>>;
+  popupId: string;
+  right?: number;
 }
