+++ frontends/article/components/ArticleActions/index.ts
@@ -1,1 +1,2 @@
 export * from './article-actions';
+export * from './article-actions-icons';
