+++ frontends/article/components/ArticleSharePopup/article-share-popup.types.ts
@@ -1,6 +1,8 @@
 import React from 'react';
 
 export interface IArticleSharePopupProps {
+  id: string;
   visible: boolean;
   setVisible: React.Dispatch<React.SetStateAction<boolean>>;
+  right?: number;
 }
